refactor wallet test
